+++ docs/src/components/NavBarIcon/index.js
@@ -1,8 +1,0 @@
-import React from "react";
-import * as Icon from 'react-feather';
-
-export const NavbarIcon = (prop) => { 
-    const { iconName, ...props } = prop;
-    const IconComponent = Icon[`${iconName}`];
-    return <IconComponent {...props} />
-}