+++ docs/src/utils/githubProfile.js
@@ -1,9 +1,0 @@
-export const githubProfile = function(id) {
-    if (id.indexOf('(') > -1) {
-        id = id.substring(id.lastIndexOf('(') + 1, id.lastIndexOf(')')).trim();
-    }
-    return {
-        id,
-        profile: `https://github.com/${id}`
-    }
-}